[agent] refactor(cli): extract pipeline + restore modules with responsibility-based names
