+++ crates/gaze-cli/src/pipeline/mod.rs
@@ -1,0 +1,4 @@
+pub(crate) mod build;
+pub(crate) mod run;
+
+pub(crate) use run::{run_clean, CleanOptions};
